Minor refactor and add Timer class
