+++ inc/HeteroGraph.hpp
@@ -41,28 +41,6 @@ public:
 };
 
 
-HeteroNode::HeteroNode(const string& name):
-        name(name),
-        type('*')
-{}
-
-
-HeteroNode::HeteroNode(string& name):
-        name(name),
-        type('*')
-{}
-
-
-HeteroNode::HeteroNode(const string& name, char type):
-        name(name),
-        type(type)
-{}
-
-
-HeteroNode::HeteroNode(string& name, char type):
-        name(name),
-        type(type)
-{}
 
 
 template<class T> class HeteroGraph {
